renamed GenerelHelper
