+++ TentakelsAttacking2/UI/Elements/PopUp/private/IntCellPopUp.cpp
@@ -5,7 +5,7 @@
 
 #include "IntCellPopUp.h"
 #include "IntCell.h"
-#include "GenerelHelper.h"
+#include "HGeneral.h"
 
 void IntCellPopUp::Initialize(AppContext const& appContext,
 	Vector2 resolution) {
